// project/bbq-mall/src/types/categorymenu.ts
export type CategoryNo = {
  categoryNo: number;
  label: string;
  depth: number;
  icon: string;
  content: string;
  children: CategoryNo[];
};

export type Product = {
  optionValues: Record<string, unknown>;
  shippingArea: string;
  additionDiscountAmt: number;
  additionDiscountUnitType: string;
  maxCouponAmt: number;
  couponDiscountAmt: number;
  accumulationAmtWhenBuyConfirm: number;
  reservationData: Record<string, unknown>;
  mainBestProductYn: boolean;
  imageUrls: string[];
  stickerLabels: Record<string, unknown>;
  mainStockCnt: number;
  maxDiscountAmount: number;
  hsCode: string;
  brandNameKo: Record<string, unknown>;
  brandNameEn: Record<string, unknown>;
  brandNameType: Record<string, unknown>;
  brandName: Record<string, unknown>;
  minSalePrice: number;
  maxSalePrice: number;
  contentsIfPausing: string;
  productType: string;
  canAddToCart: boolean;
  productNameEn: string;
  saleStatusType: string;
  promotionText: Record<string, unknown>;
  likeCount: number;
  liked: boolean;
  productManagementCd: string;
  immediateDiscountStartYmdt: Record<string, unknown>;
  immediateDiscountEndYmdt: Record<string, unknown>;
  frontDisplayYn: boolean;
  urlDirectDisplayYn: boolean;
  hasCoupons: {
    product: boolean;
    brand: boolean;
    category: boolean;
    partner: boolean;
    event: boolean;
  };
  productSalePeriodType: Record<string, unknown>;
  sectionProductStartYmdt: Record<string, unknown>;
  totalReviewCount: number;
  adult: boolean;
  listImageUrls: string[];
  searchProductId: string;
  sectionProductEndYmdt: Record<string, unknown>;
  accumulationInfo: {
    amount: number;
    rewardRateOfProduct: number;
    rewardRateOfMemberBenefit: number;
  };
  deliveryConditionInfo: {
    summary: string;
    deliveryFeeRanges: Record<string, unknown>[];
    criteria: Record<string, unknown>;
    rangeSummaries: Record<string, unknown>[];
    perOrderCnt: number;
  };
  enableCoupons: boolean;
  salePeriodType: string;
  saleStartYmdt: string;
  saleEndYmdt: string;
  displayCategoryNos: string;
  stockCnt: number;
  salePrice: number;
  productNo: number;
  brandNo: number;
  productName: string;
  groupManagementCode: Record<string, unknown>;
  groupManagementCodeName: Record<string, unknown>;
  registerYmdt: string;
  customProperties: Record<string, unknown>[];
  immediateDiscountAmt: number;
  saleCnt: number;
  reviewRating: number;
  couponTag: string;
  deliveryConditionType: string;
  partnerName: string;
  immediateDiscountUnitType: string;
  stickerInfos: Record<string, unknown>[];
  isSoldOut: boolean;
};

export type Categories = {
  categoryNo: number;
  label: string;
  depth: number;
  icon: string;
  content: string;
  children: Categories[];
};